Split main api from internals
