+++ go/godispatcher/registration/generators_test.go
